Move serializationa and util content to Models
