+++ sdk/digitaltwins/azure-digitaltwins-core/src/main/java/com/azure/digitaltwins/core/models/ListModelOptions.java
@@ -1,4 +1,4 @@
-package com.azure.digitaltwins.core.util;
+package com.azure.digitaltwins.core.models;
 
 import com.azure.core.annotation.Fluent;
