+++ sdk/digitaltwins/azure-digitaltwins-core/src/main/java/com/azure/digitaltwins/core/models/RequestOptions.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.digitaltwins.core.util;
+package com.azure.digitaltwins.core.models;
 
 import com.azure.core.annotation.Fluent;
